Add Settings & First Part Refactoring to Screen + Component Logic
There are now settings in which, for example, the data can be deleted and the biometric password can be deleted.

In addition, the data protection texts have been updated and there is now a seperate impressum.

Furthermore, the Encryption Modal has now been converted into the separation of container (logic of the application) and screen (what is displayed).